libflux: support builtin connectors
Problem: all connector plugins must be built as DSOs which
can be inconvenient.

The loop:// plugin is rarely used becuase the majority of unit tests
that would use it are in libflux and other code in src/common, which
is built before src/connectors.

Other connectors might be able to take advantage of global state in
libflux-core.so to their betterment, were they implemented as builtins.

Add an array of builtin connector init functions, currently empty,
and search it before we go off looking for DSOs in the file system.