libflux: check flux_op_set() _get() parameters
Problem: NULL parameters to flux_op_set() or flux_op_get()
could cause a segfault.

Fail with EINVAL If 'h' or 'option' are NULL.
Adjust checks in local and loop connectors to fit this interface.

Add unit tests that cover both the generic checks at entry
and the specific checks in connector_loop.