libflux: add builtin interthread connector
Problem: the shmem connector requires that all users awkwardly
pass in a zeromq context, which they must separately manage.

Implement a new built-in connector called "interthread".  This
connector contains a bidirectional channel with two
'struct flux_msglist' objects, which already have builtin interfaces
for pollfd/pollevents.  The channel protects reentrant access
to the message lists with individual internal mutexes.

The state that must be shared between threads in order to safely
create/connect a channel is global to libflux-core.so, protected by
mutexes.

The interthread connector is a bit more convenient to use than shmem.
In addition to discarding the zeromq context requirement, interthread
doesn't require each end to assume a bind or connect role since the
channel is allocated on first open by either end of the connection and
destroyed on last close by either end.

There are no notifications or errors generated if the peer disconnects.
Since both peers are in the same process space, this is expected to be
handled out of band, however a simple notification message or receive
error could be easily added later if helpful.

If the FLUX_OPT_ROUTER_NAME option is set on one end, request/response
route manipulation is performed on that end with the router identity
set by the option and the peer identity taken from the URI path.

Since this connector is intended to be used within a single process,
it performs no authentication on its peer, and allows message
credentials to pass through.  If credentials are not set, they
are initialized to [getuid(2), FLUX_ROLE_OWNER | FLUX_ROLE_LOCAL]
by the send handler.