libflux: add FLUX_OPT_ROUTER_NAME
Problem: code for manually managing request/reply route stacks is
duplicated in libtestutil and the broker.

Define a flux_t handle "option" that can be used to set a router
identity in connectors that implement the option.  If set, the connector
can perform route stack manipulation internally like the zeromq ROUTER
socket does.