python: fix to_dict() for when some attributes are missing
Problem: When a JobInfo is not fully populated with attributes,
e.g. when not all attrs were requested in a job-list RPC, or
when using flux.job.result(), the to_dict() method raises an
AttributeError exception because some derived attributes are
not present in the JobInfo object.

Skip attributes that raise AttributeError.