python: prevent exception with t_remaining JobInfo attribute
Problem: If the status and/or expiration properties of a JobInfo
object are not set, then getting the t_remaining attribute throws
an AttributeError, but this attribute should instead be set to 0.0
when unset.

Add a try/except block around the calculation of remaining time and
return 0.0 for this attribute if AttributeError or KeyError are caught.