broker: use interthread not shmem for modules
Problem: broker modules use the shmem connector to communicate,
but the interthread connector provides a lighter weight transport
in which message do not have to be encoded and decoded at each end.

Furthermore, using the interthread connector here opens the door
to reducing the number of times a message is copied when it traverses
the broker.

Now that zeromq sockets are not involved, drop the broker logic that
skips zmq_ctx_term() if any modules had to be canceled.  That eliminates
the error message: "skipping 0MQ shutdown due to presumed module socket
leak".

Update a test that was checking for the above message.