libflux: add FLUX_O_NOREQUEUE
Problem: it would be simpler to have broker modules communicate
with back to back flux_t handles, but each one adds a file descriptor
for the "requeue" queue, and one for the epoll fd used to wrap
the queue and the connector's pollfd.

To avoid an explosion in the number of used file descriptors for
this use case, add a handle flag FLUX_O_NOREQUEUE which disables
the requeue queue and causes flux_pollfd() to return the underlying
connector pollfd.

This is only appropriate to be used when messages are received without
requeuing.  flux_recv() with a match other than FLUX_MATCH_ANY, and
therefore the higher level RPC functions are not permitted to be used
directly on such a handle.