libflux: add flux_send_new ()
Problem: it's a common pattern to create a message, send it
with with flux_send(), then destroy it; however, there is no
interface to allow flux_send() to just take over the message
and thereby avoid a costly message copy if/when possible.

Add flux_send_new() which accepts a non-const flux_msg_t **.
The reference count must be 1, and on success *msg is set to NULL,
so there should be no possibility of the message being reused.

This just calls flux_send() internally if the connector does not
implement op->send_new(), or if other optional things in the
environment make calling op->send_new() unworkable (such as
RPC tracking which wants to temporarily hold copy of each request
until a response is received).

Move the call to profiling_msg_snapshot() to before the send rather
than after so send_new() doesn't have to fall back to send() when
flux is built with caliper profiling.