flux_send_new(3): add new manpage stub
Problem: there is no man page for flux_send_new().

Add it to flux_send.rst and configure a stub to be produced for
flux_send_new().

The previous synopsis function prototypes were not within a literal
block and had to be escaped.  Fix that while adding the new function.