libflux: split flux_requeue() into two functions
Problem: flux_requeue() with a 'flags' argument that determines
whether the message goes on the front or the back of the queue
is poor API design.

Better to just have two separate functions.

Drop the flags argument from flux_requeue() and have it push to
the front.  Add flux_enqueue() which pushes to the back.

Update unit tests and users.