libflux: add private handle requeue functions
Problem: flux_requeue() is not needed outside of flux-core
and has an awkward function signature.

Create two new non-public functions to replace flux_requeue()
- handle_requeue_push_front()
- handle_requeue_push_back()

Move the internal handle struct into handle_private.h so that
handle.c can be split into multiple source files.

Define the new functions in handle_requeue.h so the definitions
are not included in the installed <flux/core.h>.

Implement the functions in handle_requeue.c so that in-tree users
can statically link with the corresponding object without pulling
in all of the dependencies of handle.o.

Switch libflux users and unit tests to the new functions.

Mark flux_requeue() as deprecated.