testsuite: remove job shell standalone tests
Problem: There exist a set of job shell standalone tests that use
the --standalone option of flux-shell(1), but this option will soon
be removed.

Remove the job shell standalone tests. They appear to mostly be
duplicates of other, non-standalone tests in the testsuite anyway.

Move one test (a test for invalid arg handling) to the main job shell
sharness test.