testsuite: avoid shell standalone mode in initrc tests
Problem: The tests in t2603-job-shell-initrc.t use the job shell
--standalone option, but that option is going away.

Convert tests to use `flux run` to launch the job shell.

Remove the `setopt` test which ensured that flux_shell_setopt(3) works
when the jobspec options object is not set. It is not possible to submit
a jobspec and load an alternate initrc without using the shell options
object, so this test can no longer be maintained (and it is unlikely
that the condition will come up in the real world, so the test was of
limited value).