build: improve libflux build efficiency
Problem: running a 'make' in libflux is slow.

The slowness seems to result from the way we are defining many C
preprocessor symbols on the command line.

Modify configure.ac to define some of those in config.h and do the
rest in the C source file.

Serial build times on my system are reduced from 17.3s to 3.3s.

Defining CPPFLAGS separately for libflux.la was causing every object
to be prefixed with the library name.  Now that there are fewer flags,
just add them to AM_CPPFLAGS.