idset_create(3): improve complexity description
Problem: although the opening paragraph gives big O for a list of
vEB operations, it may not be very clear how those operations map
onto the functions in the API.

Rewrite that description using better operation names.
Mention that first and last complete in constant time.
Improve formatting of the HTML.