job-manager: allow duration update of running job
Problem: Adjusting the duration of a running job is currently not
supported, but this would be useful for jobs that are about to exceed
their time limit but have not yet computed the desired result.

Allow updates of duration for running jobs by the instance owner only
in the update-duration plugin. Since the update is being requested
by the instance owner, no validation is required and the new duration
can be set to anything, including unlimited.

Adjust the error message when a user attempts to update the duration
of one of their own running jobs. Fix one test that expects the old
error message.