testsuite: move test for update of running job
Problem: In t2290-job-update.t a test for the update of a running
job is expected to fail, but soon this operation will be supported.

Move the test to after the update-duration plugin is reloaded with
owner_allow_any=0. This will allow the test to continue to pass
even after the instance owner is allowed to update the duration of
a running job.