sched-simple: add sched.expiration RPC handler to sched-simple
Problem: RFC 27 defines a sched.expiration RPC allowing the job
manager to request an update to the expiration of a running job,
but there is no way to test operation of this RPC.

Add a handler for the sched.expiration RPC. Unpack the expected
arguments for a sanity check and return success in most cases. Add
a module debug parameter that allows the RPC to result in an error
for testing purposes.