man1: enable HTML cross-referencing of options
Problem: the HTML cross-reference tags that are automatically
generated for declared options point to a path that is relative
to the document basename name rather than the command or sub-command
name, which could be ambiguous.

For example, consider the flux job attach "-l, --label-io" option.
The auto-generated tag is "flux-job.html#cmdoption-l"

As noted in [1], the 'program' directive notifies Sphinx that all
following option directives document options for the named program.
After declaring the program to be "flux job attach", the tag becomes:
"flux-job.html#cmdoption-flux-job-attach-l".

References to the option with the program name in it like
:option:`flux job attach -l` or :option:`flux job attach --label-io`
become hyperlinks to the option description within the page containing
the program.

Note that a program directive remains in effect when included text is
processed.  This means that options declared in common/*.rst are properly
tagged when included from other pages.  For example, the "-N, --nodes=N"
option in flux-batch(1) is "flux-batch.html#cmdoption-flux-batch-N",
while in flux-run(1) it is "flux-run.html#cmdoption-flux-run-N".

[1]https://www.sphinx-doc.org/en/master/usage/domains/standard.html