man1: make command options pretty in HTML, x-ref
Problem: when referring to options, we usually just quote them
as a literal string, which is ugly in the HTML rendering
and does not allow sphinx to cross-reference references to
descriptions.

Use the :option: directive to tie option references to their
description in the HTML.  This also improves the HTML rendering.

The cross references are particularly helpful in the job submission
command options, where there are many options and some need to refer
to others.

It should be OK to use the :option: directive with any option string
since, at a minimum, it helps unify their presentation.  However Sphinx
only cross-references certain option string formats.  Consider this
example:

.. program:: flux run

.. option:: -o, --option=OPTS

The following are cross-referenced:

:option:`-o`
:option:`--option`
:option:`flux run --option`

The following are not cross-referenced by sphinx 3.4.3, but the
documentation implies the first two should work in sphinx 5.3:

:option:`--option=OPTS`
:option:`--option=pty.interactive`
:option:`-o, --option`

In this commit, all of the above are deemed acceptable markup,
athough in some cases the last format was converted to the long
option alone when it wasn't the only documentation of the short
option.

This commit only addresses option formatting, although some lines
were re-broken in the .rst files to avoid going over 79 columns.

Note: this commit is not a complete conversion of all option
strings across section 1. It only touches some of the more widely
used and/or easy to convert pages.