job-manager: update running jobs if necessary after R update
Problem: Jobs that do not specify an explicit duration may have their
expiration automatically set to the expiration of the current resource
set as seen by the scheduler. If an expiration update is then issued
for the enclosing instance, these running jobs will be terminated
before the new instance expiration.

Add a watch of resource.R to the job manager update service. If an
update occurs that modifies the resource set expiration, then walk
the list of running jobs and update any job's expiration that does not
have an explicit duration set in jobspec, but does have an expiration
set in R. This implies that the expiration was applied automatically
by the scheduler and thus the expiration should be extended.