testsuite: add rpc scaling test
Problem: no test is available to probe the performance impacts
of low level message/rpc changes.

Add a unit test called 'rpcscale'.  The test sends a batch of RPCs
to a message handler in another thread, then lets the reactor run
to process all the responses in a batch.  Both batches are timed.

When run by 'make check' with no options, the test sends enough RPCs
to cause the tagpool to expand but completes quickly.  Options allow
various things to be tried.  For example, to test 10 iterations with
a batch of 1M RPCs use

   ./test_rpcscale.t --count=1000000 --iter=10

Other options enable various flags to be set.