job-exec: add resoure_set_update_expiration()
Problem: The job-exec resource_set class does not have a method
to update the expiration of an existing resource set object, but
this will be necessary to handle resource expiration updates from
the job manager.

Add resource_set_update_expiration().

Add a sanity check of this interface to the rset unit tests.