job-exec: add handler for exec.expiration RPC
Problem: The job manager will begin sending job expiration updates
to the job execution system via a new exec.expiration RPC, but the
job-exec module does not currently support this RPC.

Add a handler for the expiration update request. Update the expiration
in the job's resource set and reset the expiration timer on update.