job-manager: send expiration update to execution system
Problem: The job execution system is unaware of updates to a job's
expiration time after a resource-update event modifies it.

On a resource-update event, send the updated expiration to the job
execution service. If the RPC fails, post a non-fatal exception to
the job's eventlog so that there is a trace of why the job may have
been terminated early or late.