job-exec: add debug flag to fail exec.expiration RPC
Problem: Testing an error response from the exec.expiration RPC from
the job manager to the job-exec module is impossible, but it would
be valuable to test this corner case.

Add a debug flag to the job-exec module which, when set, causes the
the module to send an error response to the exec.expiration RPC.
This can be used in testing to cover this case.