flux-environment(7): add new man page
Problem: environment variables are not well documented.

Add a new man page that attempts to list all the environment
variables, using the "envvar" directive so that they can be
cross referenced in the HTML rendering of other man pages.

Change the section 7 index to use a glob (like section 1) rather
than an explicit list so we don't have to manually put pages
in order and remember to update the index when pages are added.