man: make environment var pretty/x-ref
Problem: environment variables are usually tagged as literals
which causes the HTML to render in red monospace, without cross
referercing.

Use the :envvar: role to format environment variables.  If the variable
was declared with the envvar directive, then a cross-reference is
generated in HTML.  Since most of the environment variables are now
declared that way in flux-environment(7), this is usually the case.