flux-proxy: fix double-free on lost connection
Problem: flux-proxy core dumps when It loses its connection
to Flux.

ctx.remote_uri_authority is free()ed twice when flux_reactor_run()
returns < 0, as it does when the connection is lost.  Pull that free()
out of proxy_command_destroy_usock_and_router() which is called twice
and is otherwise idempotent and place it after the second call.

Fixes #5528