libjob: check security.owner when signing jobs
Problem: flux_job_submit() does not work over flux-proxy(1)
for guest users.

The peercred method of determining the instance owner, used in
flux_job_submit(), only works when the local connector is a
direct link between submission client and broker.  With flux-proxy(1)
the local connector links the submission client and flux-proxy,
which is always running as the local user, so the check in
flux_job_submit() erroneously concludes that sign=none can be used.

Change flux_job_submit() to fetch the security.owner broker attribute
and check that instead.  This incurs a small overhead on first use for
the synchronous RPC to the broker, but after that the value is cached.