mpi: drop intel_mpi shell plugin
Problem: the intel_mpi shell plugin sets I_MPI_PMI_LIBRARY to
point to the flux libpmi.so, but only if already set by the user.

This was probably done to avoid I_MPI_PMI_LIBRARY leaking through
from Slurm when flux is launched by slurm, but this method may
be surprising to a user that sets the variable on purpose.

Drop the intel_mpi shell plugin since this is all it does.

Add I_MPI_PMI_LIBRARY to the broker blocklist.  If Slurm sets it
in Flux's environment, it won't leak through to the initial program,
where it could find its way into job environments.

Update MPI personalities sharness test.
Update flux-environment(7)