pmi: set FLUX_PMI_LIBRARY_PATH only in job environ
Problem: the flux(1) command driver sets FLUX_PMI_LIBRARY_PATH,
and the broker stores it in an attribute, but only pre-v5 OpenMPI
jobs need it, and it should not be set when the PMI-1 server
required by libpmi.so is disabled.

Set FLUX_PMI_LIBRARY_PATH in the shell pmi plugin.

Also:
- Don't set FLUX_PMI_LIBRARY_PATH in flux(1).
- Don't set the conf.pmi_library_path broker attribute.
- Add FLUX_PMI_LIBRARY_PATH to the broker blocklist.
- Update the flux-config(1) explanation of pmi_library_path builtin.
- Update flux-environment(7) description of FLUX_PMI_LIBRARY_PATH.
- Drop flux-broker-attributes(7) description of conf.pmi_library_path.
- Update sharness test.