flux: don't add parent's FLUX_MODULE_PATH
Problem: flux --parent fetches the parent broker's conf.module_path
attribute and prepends it to FLUX_MODULE_PATH for subcommands, but
this has no effect.

flux-module(1) used to search FLUX_MODULE_PATH for modules and pass
the result to the broker, but recently that changed so that the broker
performs its own search.  Setting FLUX_MODULE_PATH in the flux-module
command's environment is not required since the parent will always use
its own path.

Drop this code.