job-manager: fix bad error message on duplicate plugin load
Problem: The error message recieved by a client trying to load a
duplicate jobtap plugin is "plugin not found" instead of the expected
(and more helpful) "already loaded". This occurs because errno is not
set when a plugin with the same name is found in the plugin hash,
and so a previous errno value of ENOENT is returned to the calling
function.

Set errno to EEXIST when a duplicate is found in the plugin hash.