flux-submit: fix failure to render {cc} for cc=0
Problem: When using `flux submit -cc=0-1 ..` the command should allow
substituion of `{cc}` with the current value of cc, but this fails
for cc=0 and a literal `{cc}` ends up substituted, e.g.

 $ flux submit --quiet --label-io --cc=0-1 -N1 -n1 --watch echo {cc}
 f6tye4TnY: 0: 1
 f6tye4TnX: 0: {cc}

The issue is that a value of 0 for cc evaluates as False in the test
for whether `{cc}` should be substituted. Use `isinstace(i, int)`
instead of just `i` in the test for whether substitution is required.

Fixes #5539