libsubprocess: avoid short writes in flux_subprocess_write()
Problem: The flux_subprocess_write() function may return a short write
if all the data being written does not fit in the destination buffer,
but short writes are tricky to handle and mostly cause a fatal error
in current code anyway. In some places, short writes are not detected
and would result in unpredictable behavior.

Avoid short writes in flux_subprocess_write() by checking the available
buffer space before writing. Return -1 with errno set to ENOSPC if
there is not enough available space to write the entire buffer.