libsubprocess: test ENOSPC error from flux_subprocess_write()
Problem: No tests in the libsubprocess unit tests ensure that
flux_subprocess_write() returns -1 wih errno == ENOSPC when the
requested data cannot be buffered.

Add a test that creates a subprocess with a very small buffer and
ensures that a write exceeding that buffer space returns -1 with ENOSPC
(instead of a short write).