job-ingest: configure pipeline after conf update is processed
Problem: When a configuration update is processed in job-ingest,
pipeline_configure() is run before the job-ingest module iself reads
and processes the config. However, some parameters in the top level
ingest configuration may have to be passed to pipeline_configure()
in the future, so it would be better to configure the pipeline after
reading config.

Move pipeline_configure() to the end of job_ingest_configure().