job-ingest: set default worker input buffer to 10MB
Problem: The input buffer for job-ingest pipeline workers uses the
libsubprocess default, which is currently 4MB. This default may be
too small to handle rapid submission of jobspec of moderate size,
which could result in errors at submission time.

Extend the default buffer for ingest workers to 10MB to make it less
likely for the buffers to fill up during job submission.