flux-shell(1): rename SHELL OPTIONS section
Problem: the job submission commands have SHELL OPTIONS sections
that duplicate info in the JOBSPEC OPTIONS section of flux-shell(1).

Rename JOBSPEC OPTIONS to SHELL OPTIONS, so that it will be more
recongnizable that the described options are the same ones as in
SHELL OPTIONS in the other pages.

Fold one paragraph that was inline with the options into the section
introduction.

Add cross-reference tags for SHELL OPTIONS and SHELL INITRC.