doc: add a note about Python module path to flux-environment(7)
Problem: The Python module search path is modified at startup by
Flux Python utilities, but this is not documented anywhere, which
could cause confusion when user additions to PYTHONPATH are modified
or ignored.

Add a note to flux-environment(7) about how `sys.path` is modified
at subcommand startup, and what to do if (unlikely) this behavior
needs to be worked around.