testsuite: test validator handling of modified PYTHONPATH
Problem: There is no reproducer for issue #5547 - user modification
of PYTHONPATH causes job validator to fail.

Add a reproducer to t2110-job-ingest-validator.t. Use a dummy urrlib
module to reproduce the issue and ensure validator still works when
the path to this module is first in PYTHONPATH.

This issue actually affects all Python commands and any Python module
that might be placed first in PYTHONPATH, but the job validator and
urllib are used in this test as a surrogate since it reproduces the
original issue as reported.