cmd: py-runner.py: add wrapper script for Python commands
Problem:  If a user modifies PYTHONPATH, this could cause incompatible
versions of standard or system installed Python modules to be loaded
by Flux Python commands, causing them to fail in unpredictable ways.
Currently, there is no way to prevent this without running some code
in all Python commands before the first module import.

Add a new script that can be used as a wrapper for Flux Python
utilities and which modifies sys.path before running the script with
runpy.run_path(). Paths are pushed to the front of sys.path such that
the following are guaranteed to be at the front:

 - FLUX_PYTHON_PATH_PREPEND
 - Flux builtin module search path
 - Current interpreter standard paths in the following order:
   - stdlib
   - platform stdlib
   - platform module path
   - pure python module path

This works around the issue of a path with invalid or incompatible
modules first in the path, but just for commands invoked under the
wrapper script, and does not modify the environment which could be
inherited by a user environment under job submission commands.