libflux: add path to Python command wrapper to builtin conf
Problem: The location of the Python command wrapper script should
be easy for flux(1) to determine, but currently there is no builtin
configuration holding that path.

Add a `python_wrapper` builtin config value that holds the path to
the wrapper script and can be used by `flux(1)` to fetch the value
with minimal fuss.