broker: use interthread connector internally
Problem: the broker has a strange flux_t handle configuration
dating back to early days of Flux before we had nice things.

It is an exposed handle implementation with only a send() operation.
Instead of queuing messages, send() directly passes messages to routing
logic.  When routing logic wants to send a message back to the handle, it
uses flux_requeue (RQ_TAIL).

Replace that with back to back interthread flux_t handles. This more
conventional setup should be less fragile and easier to reason about.

One material change is that messages sent on the handle (e.g. when some
part of the broker makes an RPC) are now queued and later dequeued by
a reactor handler and routed, rather than immediately routed.  As a
consequence, synchronous RPCs on the broker handle cannot work at all.