libmissing: define __W_EXITCODE() if missing
Problem: __W_EXITCODE() is missing on Alpine linux, which causes
compilation to fail in various source modules.

Add a fallback definition of __W_EXITCODE() in libmissing/macros.h
and include it from source modules that require this macro.

The macro was borrowed from glibc.  Although the bitfields in the
wait(2) status byte do not appear to be standardized by POSIX,
this comment in the glibc source indicates that their implementation
may be fairly portable:

/* Everything extant so far uses these same bits.  */