libflux: improve perf of flux_msglist_disconnect/cancel
Problem: The flux_msglist_disconnect() and flux_msglist_cancel()
functions repeat a series of non-trivial operations on the
source message for each target message in the list on which they
are attempting the match. For large message lists, this can cause
these functios to take much longer than required.

Add a new flux_msg_match struct to disconnect.h and a function to
initialize this structure with the message attributes that will be
used in flux_msglist_disconnect() and flux_msglist_match(). Add new
_ex versions of the functions that take a match argument instead of
a message so that the data only needs to be obtained once.
Expose the _ex functions in the header so that users that aren't able
to use a msglist can still take advantage of the more efficient
matching.

Update flux_msglist_disconnect() and flux_msglist_cancel() to use
the _ex versions of the functions.