testsuite: skip test if pipe size is too small
Problem: the librouter sendfd unit test hangs on Alpine linux.

One of the tests writes an 8K+ message one end of a pipe without a
reader on the other end, and expects it not to block.

Check the pipe size with fcntl(2) and adjust it upward if necessary.
If that fails, skip those tests.

Note: the pipe size was observed to be exactly 8K in the failing
environment.  However, other Alpine environments had a larger pipe
size.  This was not explained except to note that Flux was built
from spack on in the failing case.