replace %F in str[pf]time() with %Y-%m-%d
Problem: the %F format specifier in strptime(3) and strftime(3)
does not work on Alpine linux, causing some unit tests to fail.

The %F format specifier is a non-POSIX alias for %Y-%m-%d that is
implemented by glibc but not by musl libc.

Replace the %F alias with the full format %Y-%m-%d wherever it occurs.