ci: add --quick-check option to docker-run-checks.sh
Problem: In may be useful to run a quick build + confidence check
in CI instead of the full testsuite. This is especially the case when
porting to new systems where the entire set of sharness tests may
not be working yet.

There is already a --unit-tests-only option, but this doesn't build
any sources under ./t, and it would be nice to include that along
with a simple verification check like `flux start flux run hostname`.

Add a --quick-check option to docker-run-checks.sh that completes a
`make check-prep` (which builds all sources and runs unit tests),
followed by a confidence check that starts a 2 broker test instance
and runs a series of jobs.