ci: add alpine build to github ci
Problem: There is no automated build of flux-core on a musl
libc system, so PRs are not checked for portability to non-glibc
environments.

Add a ci build on alpine to the github workflow matrix. Only run a
quick check since there are still many fixes required for sharness
tests in this environment.