kvs-watch: improve performance of watcher_cancel_all
Problem: Profiling indicates that watcher_cancel_all() in the kvs-watch
module is taking more time than expected.

This hotspot appears because flux_cancel_mach() is used to compare
the request message to the list of current watchers and this function
unpacks the matchtag from the request on each match, therefore the
request message is unpacked many times unnecessarily.

Unpack the matchtag from the request once and pass it down through
watcher_cancel_ns() and watcher_cancel() so the value can be reused.