job-info: send ENODATA only on cancel
Problem: In guest-watch and watch code, ENODATA is sent to the original
caller both when the stream is canceled and when it is disconnected.  It
should not be sent when it is disconnected.

Add an internal flag to only send ENODATA when the watch stream is
canceled, not when it is disconnected.