flux-kvs(1): put sub-commands in COMMANDS subsects
Problem: flux-kvs(1) sub-commands are not demarcated by
document subsections, and the options for each sub-command
are not properly formatted.

Copy the structure of git-remote(1) and place subcommands in
sub-sections of a COMMANDS section, with the subcommand name as
the sub-section title in lower case.

Declare the program name in each sub-section for x-ref, and
define options properly so they are formatted in the HTML rendering
like the other man pages.

Option descriptions were altered to fit the new format, and reworded
in some cases.