flux-broker: create rundir/bin
Problem: the ssh connector cannot always find flux-relay
on a remote system.

Since the ssh connector does have the path to the remote socket
and hence the remote broker rundir, create a bin directory in the rundir
that includes a flux symlink pionting to the instance's flux(1).
The ssh connector can prepend the bin directory to the remote PATH and
thus find flux-relay more reliably.